requirements.txt
# Core dependencies
python-dotenv>=1.0.0
openai>=1.0.0
flask>=3.0.0
discord.py>=2.3.0
Pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
gunicorn>=21.0.0

# Optional: for better text rendering
# pillow-heif>=0.13.0
